new things about context

+++ projeto-final/backend/resolvers/Mutation/usuario.js
@@ -13,7 +13,9 @@ const mutations = {
             }
         })
     },
-    async novoUsuario(_, { dados }) {
+    async novoUsuario(_, { dados }, context) {
+        context && context.validarAdmin()
+
         try {
             const idsPerfis = []
 
@@ -49,7 +51,9 @@ const mutations = {
             throw new Error(e.sqlMessage)
         }
     },
-    async excluirUsuario(_, args) {
+    async excluirUsuario(_, args, context) {
+        context && context.validarAdmin()
+
         try {
             const usuario = await obterUsuario(_, args)
             if(usuario) {
@@ -65,12 +69,14 @@ const mutations = {
         }
 
     },
-    async alterarUsuario(_, { filtro, dados }) {
+    async alterarUsuario(_, { filtro, dados }, context) {
+        context && context.validarUsuarioFiltro(filtro)
+
         try {
             const usuario = await obterUsuario(_, { filtro })
             if(usuario) {
                 const { id } = usuario
-                if(dados.perfis) {
+                if(context.admin && dados.perfis) {
                     await db('usuarios_perfis')
                         .where({ usuario_id: id }).delete()
